+++ crates/biome_js_analyze/tests/specs/nursery/noMisrefactoredShorthandAssign/invalid.js
@@ -1,0 +1,34 @@
+a += (a + b)
+
+object.a += object.a + b;
+
+a -= a - b
+
+a *= a * b
+
+a *= b * a
+
+a /= a / b
+
+a %= a % b
+
+a **= a ** b
+
+a >>= a >> b
+
+a <<= a << b
+
+a >>>= a >>> b
+
+a &= a & b
+
+a &= b & a
+
+a |= a | b
+
+a |= b | a
+
+a ^= a ^ b
+
+a ^= b ^ a
+