+++ crates/biome_service/src/configuration/parse/json/rules.rs
@@ -2034,6 +2034,7 @@ impl VisitNode<JsonLanguage> for Nursery {
                 "noGlobalIsNan",
                 "noInvalidNewBuiltin",
                 "noMisleadingInstantiator",
+                "noMisrefactoredShorthandAssign",
                 "noUnusedImports",
                 "noUselessElse",
                 "noVoid",
@@ -2310,6 +2311,29 @@ impl VisitNode<JsonLanguage> for Nursery {
                         diagnostics,
                     )?;
                     self.no_misleading_instantiator = Some(rule_configuration);
+                }
+                _ => {
+                    diagnostics.push(DeserializationDiagnostic::new_incorrect_type(
+                        "object or string",
+                        value.range(),
+                    ));
+                }
+            },
+            "noMisrefactoredShorthandAssign" => match value {
+                AnyJsonValue::JsonStringValue(_) => {
+                    let mut configuration = RuleConfiguration::default();
+                    self.map_to_known_string(&value, name_text, &mut configuration, diagnostics)?;
+                    self.no_misrefactored_shorthand_assign = Some(configuration);
+                }
+                AnyJsonValue::JsonObjectValue(_) => {
+                    let mut rule_configuration = RuleConfiguration::default();
+                    rule_configuration.map_rule_configuration(
+                        &value,
+                        name_text,
+                        "noMisrefactoredShorthandAssign",
+                        diagnostics,
+                    )?;
+                    self.no_misrefactored_shorthand_assign = Some(rule_configuration);
                 }
                 _ => {
                     diagnostics.push(DeserializationDiagnostic::new_incorrect_type(